updates
Face orientation - get the euler angle of the face using the values of the mtcnn landmarks (mouth, nose, eyes), and also, using opencv's solvePnP or perspective-n-point by projecting a 3d point onto the image plane to calculate pitch, yaw and roll angles of the face.

eyes and mouth open/close detector - using the mediapipe's facemesh, we first plot the left and right eye indices and check if the eyes/mouth is open by comparing it to a closed ratio through euclidean. and by top/bottom/mouth heigh for the mouth.